Example Template for Build Report
This is simple test that mimics the beginning of the
output generated by `Report2Html.tcl`. I haven't included
everything, my purpose is only to show what's possible.

Run with:

```
tclsh ReportTemplate.tcl > report.html
```

This will generate a file `report.html` that can be viewed in a
browser window.